scripts/flash.sh: erase userdata/misc/metadata as the last step

+++ scripts/flash.sh
@@ -56,14 +56,6 @@ fastboot flash vbmeta_vendor_dlkm_a $OUT/vbmeta_vendor_dlkm.img $disable_verity
 if ! [ $? -eq 0 ] ; then echo "Failed to flash vbmeta_vendor_dlkm"; exit 1; fi
 fastboot flash super $OUT/super_empty.img
 if ! [ $? -eq 0 ] ; then echo "Failed to flash super"; exit 1; fi
-if ! [ ${skip_userdata} -eq 1 ] ; then
-	fastboot erase userdata
-	if ! [ $? -eq 0 ] ; then echo "Failed to erase userdata"; exit 1; fi
-	fastboot erase metadata
-	if ! [ $? -eq 0 ] ; then echo "Failed to erase metadata"; exit 1; fi
-	fastboot erase misc
-	if ! [ $? -eq 0 ] ; then echo "Failed to erase misc"; exit 1; fi
-fi
 fastboot reboot fastboot
 fastboot wait-for-device
 fastboot flash system_a $OUT/system.img
@@ -72,4 +64,13 @@ fastboot flash vendor_a $OUT/vendor.img
 if ! [ $? -eq 0 ] ; then echo "Failed to flash vendor"; exit 1; fi
 fastboot flash vendor_dlkm_a $OUT/vendor_dlkm.img
 if ! [ $? -eq 0 ] ; then echo "Failed to flash vendor_dlkm"; exit 1; fi
+if ! [ ${skip_userdata} -eq 1 ] ; then
+	fastboot reboot bootloader
+	fastboot erase userdata
+	if ! [ $? -eq 0 ] ; then echo "Failed to erase userdata"; exit 1; fi
+	fastboot erase metadata
+	if ! [ $? -eq 0 ] ; then echo "Failed to erase metadata"; exit 1; fi
+	fastboot erase misc
+	if ! [ $? -eq 0 ] ; then echo "Failed to erase misc"; exit 1; fi
+fi
 fastboot reboot
